refactor: move base indicator classes to own dir
makes it easier to see the concrete classes.

also renamed XIndicators -> XIndicatorGroup to prep for XIndicator
being an unambiguous component name.
