+++ src/web/topic/components/Indicator/Base/Indicator.styles.tsx
