+++ src/web/topic/components/Indicator/NotesIndicator.tsx
@@ -3,7 +3,7 @@ import { ButtonProps } from "@mui/material";
 import { useCallback } from "react";
 
 import { emitter } from "@/web/common/event";
-import { Indicator } from "@/web/topic/components/Indicator/Indicator";
+import { Indicator } from "@/web/topic/components/Indicator/Base/Indicator";
 import { setSelected } from "@/web/view/selectedPartStore";
 
 interface Props {
